Refactors to move LocalFeedLoader and FeedStore collaborator to their own file in the production module.
