+++ FeedModule/FeedModule/FeedCache/FeedStore.swift
@@ -1,0 +1,16 @@
+//
+//  FeedStore.swift
+//  FeedModule
+//
+//  Created by Omran Khoja on 2/9/22.
+//
+
+import Foundation
+
+public protocol FeedStore {
+    typealias DeletionCompletion = (Error?) -> Void
+    typealias InsertionCompletion = (Error?) -> Void
+    
+    func deleteCachedFeed(completion: @escaping DeletionCompletion)
+    func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
+}
