+++ FeedModule/FeedModule/FeedCache/LocalFeedLoader.swift
@@ -1,0 +1,37 @@
+//
+//  LocalFeedLoader.swift
+//  FeedModule
+//
+//  Created by Omran Khoja on 2/9/22.
+//
+
+import Foundation
+
+public final class LocalFeedLoader {
+    private let store: FeedStore
+    private let currentDate: () -> Date
+    
+    public init(store: FeedStore, currentDate: @escaping () -> Date) {
+        self.store = store
+        self.currentDate = currentDate
+    }
+    
+    public func save(_ items: [FeedItem], completion: @escaping (Error?) -> Void) {
+        store.deleteCachedFeed { [weak self] error in
+            guard let self = self else { return }
+            
+            if let cacheDeletionError = error {
+                completion(cacheDeletionError)
+            } else {
+                self.cache(items, with: completion)
+            }
+        }
+    }
+    
+    private func cache(_ items: [FeedItem], with completion: @escaping (Error?) -> Void) {
+        store.insert(items, timestamp: self.currentDate()) { [weak self] error in
+            guard self != nil else { return }
+            completion(error)
+        }
+    }
+}
